Move src/stories to stories/src
