+++ stories/src/AnchoredOverlay.stories.tsx
