+++ stories/src/AvatarStack.stories.tsx
