+++ stories/src/IssueLabelToken.stories.tsx
