+++ stories/src/Label2.stories.tsx
