+++ stories/src/TextInputWithTokens.stories.tsx
