+++ stories/src/useAnchoredPosition.stories.tsx
